requirements.txt
tensorflow
keras
streamlit
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy